% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_options.R
\name{annotation_options}
\alias{annotation_options}
\title{List available annotation options for a given organism}
\usage{
annotation_options(organism = c("human", "mouse"))
}
\arguments{
\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}
}
\value{
A \code{character()} vector with the supported annotation options for the
given \code{organism}.
}
\description{
This function will return the available annotation options for a given
\code{organism}.
}
\examples{

annotation_options("human")
annotation_options("mouse")
}
