% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_locate_url.R
\name{file_locate_url}
\alias{file_locate_url}
\title{Title}
\usage{
file_locate_url(
  project,
  project_home = project_home_available(type = type, organism = organism, recount3_url
    = recount3_url),
  type = c("metadata", "gene", "exon", "jxn", "bw"),
  organism = c("human", "mouse"),
  sample = NULL,
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)
}
\arguments{
\item{project}{A \code{character(1)}}

\item{recount3}{}
}
\value{

}
\description{
Title
}
\examples{

file_locate_url("ERP001942", "sra", "metadata", "human")

file_locate_url("ERP001942", "sra", "bw", "human", "ERR204900")
}
