% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_locate_url_annotation.R
\name{file_locate_url_annotation}
\alias{file_locate_url_annotation}
\title{Construct the URL to a recount3 annotation file}
\usage{
file_locate_url_annotation(
  type = c("gene", "exon"),
  organism = c("human", "mouse"),
  annotation = annotation_options(organism),
  recount3_url = "http://idies.jhu.edu/recount3/data"
)
}
\arguments{
\item{type}{A \code{character(1)} specifying whether you want to access gene
counts or exon data.}

\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{annotation}{A \code{character(1)} specifying which annotation you want to
download. Only used when \code{type} is either \code{gene} or \code{exon}.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}
}
\value{
A \code{character(1)} with the URL (or file path) to access the
recount3 annotation file.
}
\description{
Given a expression feature \code{type}, \code{organism} and \code{annotation}, this
function constructs the URL (or file path) to access a recount3
annotation file. This function is used by \code{create_rse_manual()}.
}
\examples{

file_locate_url_annotation()
file_locate_url_annotation(organism = "mouse")
}
\seealso{
Other internal functions for accessing the recount3 data: 
\code{\link{annotation_ext}()},
\code{\link{create_rse_manual}()},
\code{\link{file_locate_url}()},
\code{\link{file_retrieve}()},
\code{\link{project_homes}()},
\code{\link{read_counts}()},
\code{\link{read_metadata}()}
}
\concept{internal functions for accessing the recount3 data}
