% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Temporary function}
\usage{
foo(x)
}
\arguments{
\item{x}{Any object}
}
\value{
The same object.
}
\description{
I made this function only to get around the issue at
https://github.com/r-lib/covr/issues/427
}
\examples{

foo(3)

}
