% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_home_available.R
\name{project_home_available}
\alias{project_home_available}
\title{Find available project home options}
\usage{
project_home_available(
  type = c("metadata", "gene", "exon", "jxn", "bw"),
  organism = c("human", "mouse"),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3",
  bfc = BiocFileCache::BiocFileCache()
)
}
\value{
A \code{character()} vector with the available \code{project_home} options.
}
\description{
Find available project home options
}
\examples{

project_home_available("metadata")
}
