% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3_cache.R
\name{recount3_cache}
\alias{recount3_cache}
\title{Specify where to cache the recount3 files}
\usage{
recount3_cache(dir = getOption("recount3_cache", NULL))
}
\arguments{
\item{dir}{A \code{character(1)} specifying the directory that will be used
for caching the data.}
}
\value{
A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the recount3 files will be cached to.
}
\description{
This function allows you to specify where the recount3 files will be cached
to. It is powered by
\link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}.
}
\examples{

## Locate the recount3 cache default directory
recount3_cache()
}
\seealso{
Other recount3 cache functions: 
\code{\link{recount3_cache_files}()},
\code{\link{recount3_cache_rm}()}
}
\concept{recount3 cache functions}
