% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3_hub.R
\name{recount3_hub}
\alias{recount3_hub}
\title{Access recount3 project_info objects through ExperimentHub}
\usage{
recount3_hub(
  project,
  data_source = c("sra", "gtex", "tcga"),
  organism = c("human", "mouse"),
  eh = ExperimentHub::ExperimentHub()
)
}
\arguments{
\item{project}{A \code{character(1)} specifying the project ID you are interested
in. Could be a SRA study ID or could be a specific GTEx or TCGA tissue.}

\item{data_source}{A \code{character(1)} specifying the data source.}

\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{eh}{An \code{ExperimentHub} object
\link[ExperimentHub:ExperimentHub-class]{ExperimentHub-class}.}
}
\value{
A \code{data.frame()} with one row and three columns that contain
the information needed by \code{create_rse()} for accessing the data.
}
\description{
This function downloads from \code{ExperimentHub} the list of available
recount3 projects. However, you might be more tempted to explore the
data interactively using code like
\code{eh2 <- AnnotationHub::display(ExperimentHub::ExperimentHub())}
}
\examples{

## Locate a project of interest
project_info <- recount3_hub("SRP009615")

## For an interactive display use:
if (interactive()) {
    ## Open an interactive display
    eh2 <- AnnotationHub::display(ExperimentHub::ExperimentHub())

    ## Select one study, then hit the "send" button on the interactive
    ## display

    ## Check that you selected one study
    if (length(eh2) == 1) {
        ## Download the data using the ExperimentHub [[ syntax
        project_info <- eh2[[1]]
    }
}

## Create a RangedSummarizedExperiment object for this project
create_rse(project_info)
}
