% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_projects.R
\name{available_projects}
\alias{available_projects}
\title{List available projects in recount3}
\usage{
available_projects(
  organism = c("human", "mouse"),
  recount3_url = getOption("recount3_url", "http://idies.jhu.edu/recount3/data"),
  bfc = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}
}
\value{
A \code{data.frame()} with the project ID (\code{project}), the \code{organism}, the
\code{file_source} from where the data was accessed,
the \code{recount3} project home location (\code{project_home}), and the project
\code{project_type} that differentiates between \code{data_sources} and \code{compilations}.
}
\description{
List available projects in recount3
}
\examples{

## Find all the human projects
human_projects <- available_projects()

## Explore the results
dim(human_projects)
head(human_projects)

## How many are from a data source vs a compilation?
table(human_projects$project_type, useNA = "ifany")

## What are the unique file sources?
table(
    human_projects$file_source[human_projects$project_type == "data_sources"]
)

## Note that big projects are broken up to make them easier to access
## For example, GTEx and TCGA are broken up by tissue
head(subset(human_projects, file_source == "gtex"))
head(subset(human_projects, file_source == "tcga"))

## Find all the mouse projects
mouse_projects <- available_projects(organism = "mouse")

## Explore the results
dim(mouse_projects)
head(mouse_projects)

## How many are from a data source vs a compilation?
table(mouse_projects$project_type, useNA = "ifany")

## What are the unique file sources?
table(
    mouse_projects$file_source[mouse_projects$project_type == "data_sources"]
)
}
