% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rse.R
\name{create_rse}
\alias{create_rse}
\title{Create a recount3 RangedSummarizedExperiment gene or exon object}
\usage{
create_rse(
  project_info,
  type = c("gene", "exon", "jxn"),
  annotation = annotation_options(project_info$organism),
  bfc = BiocFileCache::BiocFileCache(),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)
}
\arguments{
\item{project_info}{A \code{data.frame()} with one row that contains the
information for the project you are interested in. You can find which
project to work on using \code{todo()}.}

\item{type}{A \code{character(1)} specifying whether you want to access gene,
exon, or exon-exon junction counts.}

\item{annotation}{A \code{character(1)} specifying which annotation you want to
download. Only used when \code{type} is either \code{gene} or \code{exon}.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
Once you have identified a project you want to work with, you can use this
function to construct a \code{recount3}
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
(RSE) object at the gene or exon expression feature level. This function will
retrieve the data, cache it, then assemble the RSE object.
}
\examples{

## Find the project you are interested in
## (temporarily build this data.frame)
proj_info <- data.frame(
    project = "ERP110066",
    project_home = "data_sources/sra",
    organism = "human"
)

## Create a RSE object at the gene level
rse_gene_ERP110066 <- create_rse(proj_info)

## Explore the resulting RSE gene object
rse_gene_ERP110066

dim(rse_gene_ERP110066)
rowRanges(rse_gene_ERP110066)
colnames(colData(rse_gene_ERP110066))

## Create a second RSE object using another annotation
rse_gene_ERP110066_gencode_v29 <- create_rse(
    proj_info,
    annotation = "gencode_v29"
)

## Doesn't work right now
# rse_gene_ERP110066_refseq <- create_rse(
#     proj_info,
#     annotation = "refseq"
# )

## Create a RSE object at the exon level
\dontrun{
rse_exon_ERP110066 <- create_rse(
    proj_info,
    type = "exon"
)

## Explore the resulting RSE exon object
rse_exon_ERP110066

dim(rse_exon_ERP110066)
rowRanges(rse_exon_ERP110066)
colData(rse_exon_ERP110066)
colnames(colData(rse_exon_ERP110066))
}
}
