% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rse_manual.R
\name{create_rse_manual}
\alias{create_rse_manual}
\title{Internal function for creating a recount3 RangedSummarizedExperiment object}
\usage{
create_rse_manual(
  project,
  project_home = project_home_available(organism = organism, recount3_url =
    recount3_url, bfc = bfc),
  type = c("gene", "exon", "jxn"),
  organism = c("human", "mouse"),
  annotation = annotation_options(organism),
  bfc = BiocFileCache::BiocFileCache(),
  jxn_format = c("ALL", "UNIQUE"),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)
}
\arguments{
\item{project}{A \code{character(1)} with the ID for a given study.}

\item{project_home}{A \code{character(1)} with the home directory for the
\code{project}. You can find these using \code{project_home_available()}.}

\item{type}{A \code{character(1)} specifying whether you want to access gene,
exon, or exon-exon junction counts.}

\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{annotation}{A \code{character(1)} specifying which annotation you want to
download. Only used when \code{type} is either \code{gene} or \code{exon}.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}

\item{jxn_format}{A \code{character(1)} specifying whether the exon-exon junction
files are derived from all the reads (\code{ALL}) or only the uniquely mapping
read counts (\code{UNIQUE}). Note that \code{UNIQUE} is only available for some
projects.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
This function is used internally by \code{create_rse()} to construct a \code{recount3}
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object that contains the base-pair coverage counts at the \code{gene} or \code{exon}
feature level for a given annotation.
}
\examples{

rse_gene_ERP110066_manual <- create_rse_manual(
    "ERP110066",
    "data_sources/sra"
)
rse_gene_ERP110066_manual

rse_gene_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq"
)
rse_gene_ERP110066_collection_manual


rse_exon_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq",
    type = "exon"
)
rse_exon_ERP110066_collection_manual

system.time(rse_jxn_ERP110066_collection_manual <- create_rse_manual(
    "ERP110066",
    "collections/geuvadis_smartseq",
    type = "jxn"
))
rse_jxn_ERP110066_collection_manual

system.time(rse_jxn_ERP001942_manual <- create_rse_manual(
    "ERP001942",
    "data_sources/sra",
    type = "jxn"
))
rse_jxn_ERP001942_manual

## Mouse example
rse_gene_SRP060340_manual <- create_rse_manual(
    "SRP060340",
    "data_sources/sra",
    organism = "mouse"
)
rse_gene_SRP060340_manual
\dontrun{
project <- "ERP110066"
project_home <- "collections/geuvadis_smartseq"

project <- "ERP001942"
project_home <- "data_sources/sra"
type <- "jxn"
organism <- "human"
annotation <- "gencode_v26"
jxn_format <- "ALL"
bfc <- BiocFileCache::BiocFileCache()
recount3_url <- "http://snaptron.cs.jhu.edu/data/temp/recount3"
}
}
\references{
\url{https://doi.org/10.12688/f1000research.12223.1} for details on the
base-pair coverage counts used in recount2 and recount3.
}
\seealso{
Other internal functions for accessing the recount3 data: 
\code{\link{annotation_ext}()},
\code{\link{file_locate_url_annotation}()},
\code{\link{file_locate_url}()},
\code{\link{file_retrieve}()},
\code{\link{project_home_available}()},
\code{\link{read_counts}()},
\code{\link{read_metadata}()}
}
\concept{internal functions for accessing the recount3 data}
