% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rse_manual.R
\name{create_rse_manual}
\alias{create_rse_manual}
\title{Internal function for creating a recount3 RangedSummarizedExperiment object}
\usage{
create_rse_manual(
  project,
  project_home = project_home_available(organism = organism, recount3_url =
    recount3_url),
  type = c("gene", "exon"),
  organism = c("human", "mouse"),
  annotation = annotation_options(organism),
  bfc = BiocFileCache::BiocFileCache(),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3"
)
}
\arguments{
\item{project}{A \code{character(1)} with the ID for a given study.}

\item{project_home}{A \code{character(1)} with the home directory for the
\code{project}. You can find these using \code{project_home_available()}.}

\item{type}{A \code{character(1)} specifying whether you want to access gene
counts or exon counts.}

\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{annotation}{A \code{character(1)} specifying which annotation you want to
download. Only used when \code{type} is either \code{gene} or \code{exon}.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
This function is used internally by \code{create_rse()} to construct a \code{recount3}
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object that contains the base-pair coverage counts at the \code{gene} or \code{exon}
feature level for a given annotation.
}
\examples{

rse_gene_ERP001942_manual <- create_rse_manual("ERP001942", "data_sources/sra")
rse_gene_ERP001942_manual
}
\references{
\url{https://doi.org/10.12688/f1000research.12223.1} for details on the
base-pair coverage counts used in recount2 and recount3.
}
