% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_list.R
\name{file_list}
\alias{file_list}
\title{Find available sub-directories for a given URL}
\usage{
file_list(url, bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{url}{A \code{character(1)} with either a URL or a local file path.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}
}
\value{
A \code{character()} vector with the available options.
}
\description{
This function lists all the available files inside a given URL. If \code{url} is
a local directory, such as when you download all the \code{recount3} files to
your computer, then this function will list all the files inside \code{url}.
}
\examples{

## List the contents in a URL
file_list("http://snaptron.cs.jhu.edu/data/temp/recount3/human/data_sources/")

## List the contents in a directory
file_list(getwd())
}
