% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_retrieve.R
\name{file_retrieve}
\alias{file_retrieve}
\title{Download a cache a remote file}
\usage{
file_retrieve(url, bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{url}{A \code{character(1)} with the file URL.}

\item{bfc}{A \code{BiocFileCache} object
\link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}.}
}
\value{
A \code{character(1)} with the path to the cached file.
}
\description{
Download a cache a remote file
}
\examples{

file_url <- file_locate_url("ERP001942", "sra", "metadata", "human")
file_local <- file_retrieve(url = file_url)
}
