% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_retrieve.R
\name{file_retrieve}
\alias{file_retrieve}
\title{Download a cache a remote file}
\usage{
file_retrieve(url, bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{url}{A \code{character(1)} with the file URL or the actual local path in
which case, it won't be cached.}

\item{bfc}{A \code{BiocFileCache} object
\link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}.}
}
\value{
A \code{character(1)} with the path to the cached file.
}
\description{
Download a cache a remote file
}
\examples{

## Download the metadata file for project ERP001942
url_ERP001942_meta <- file_locate_url("ERP001942", "data_sources/sra")
local_ERP001942_meta <- file_retrieve(url = url_ERP001942_meta)
local_ERP001942_meta
}
