% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_retrieve.R
\name{file_retrieve}
\alias{file_retrieve}
\title{Download a remote file and cache it to re-use later}
\usage{
file_retrieve(url, bfc = BiocFileCache::BiocFileCache())
}
\arguments{
\item{url}{A \code{character(1)} with the file URL or the actual local path in
which case, it won't be cached.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}
}
\value{
A \code{character(1)} with the path to the cached file.
}
\description{
Download a remote file and cache it to re-use later
}
\examples{

## Download the metadata file for project ERP001942
url_ERP001942_meta <- file_locate_url("ERP001942", "data_sources/sra")
local_ERP001942_meta <- file_retrieve(url = url_ERP001942_meta)
local_ERP001942_meta

## Download the gene counts file for project ERP001942
url_ERP001942_gene <- file_locate_url("ERP001942", "data_sources/sra", type = "gene")
local_ERP001942_gene <- file_retrieve(url = url_ERP001942_gene)
local_ERP001942_gene
}
