% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_home_available.R
\name{project_home_available}
\alias{project_home_available}
\title{Find available project home options}
\usage{
project_home_available(
  organism = c("human", "mouse"),
  recount3_url = "http://snaptron.cs.jhu.edu/data/temp/recount3",
  bfc = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{organism}{A \code{character(1)} specifying which organism you want to
download data from. Supported options are \code{"human"} or \code{"mouse"}.}

\item{recount3_url}{A \code{character(1)} specifying the home URL for \code{recount3}
or a local directory where you have mirrored \code{recount3}.}

\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to.}
}
\value{
A \code{character()} vector with the available \code{project_home} options.
}
\description{
This function finds the home for a given project (study) of interest based
on the \code{organism} and the \code{home_type}.
}
\examples{

## List the different available `project_home` options for the default
## arguments
project_home_available()
}
\seealso{
Other internal functions for accessing the recount3 data: 
\code{\link{annotation_ext}()},
\code{\link{create_rse_manual}()},
\code{\link{file_locate_url_annotation}()},
\code{\link{file_locate_url}()},
\code{\link{file_retrieve}()},
\code{\link{read_counts}()},
\code{\link{read_metadata}()}
}
\concept{internal functions for accessing the recount3 data}
