% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read a metadata file}
\usage{
read_metadata(metadata_file)
}
\arguments{
\item{metadata_file}{A \code{character(1)} with the local path to a \code{recount3}
metadata file.}
}
\value{
A \code{data.frame()} with all lower case column names for the sample
metadata.
}
\description{
This function reads in a \code{recount3} metadata file into R. You can first
locate the file using \code{file_locate_url()} then download it to your computer
using \code{file_retrieve()}.
}
\examples{

## Download the metadata file for project ERP001942
url_ERP001942_meta <- file_locate_url("ERP001942", "data_sources/sra")
local_ERP001942_meta <- file_retrieve(url = url_ERP001942_meta)

## Read the metadata
ERP001942_meta <- read_metadata(local_ERP001942_meta)

## Locate and read the GTEx bladder metadata
gtex_bladder_meta <- read_metadata(
    file_retrieve(
        file_locate_url("BLADDER", "data_sources/gtex")
    )
)

head(gtex_bladder_meta)
}
