% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3-package.R
\docType{package}
\name{recount3-package}
\alias{recount3}
\alias{recount3-package}
\title{recount3: Explore and download data from the recount3 project}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Explore and download data from the recount project available at
    https://jhubiostatistics.shinyapps.io/recount3/. Using the recount3 package
    you can download RangedSummarizedExperiment objects at the gene, exon or
    exon-exon junctions level, the raw counts, the phenotype metadata used,
    the urls to the sample coverage bigWig files. The RangedSummarizedExperiment
    objects can be used by different packages for performing differential
    expression analysis. Using data from the recount3 project you can perform
    annotation-agnostic differential expression analyses as described at 
    http://doi.org/TODO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/recount3}
  \item Report bugs at \url{https://github.com/LieberInstitute/recount3/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

}
\keyword{internal}
