% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3-package.R
\docType{package}
\name{recount3-package}
\alias{recount3}
\alias{recount3-package}
\title{recount3: Explore and download data from the recount3 project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The recount3 package enables access to a large amount of uniformly processed RNA-seq data from human and mouse. You can download RangedSummarizedExperiment objects at the gene, exon or exon-exon junctions level with sample metadata and QC statistics. In addition we provide access to sample coverage BigWig files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/recount3}
  \item Report bugs at \url{https://github.com/LieberInstitute/recount3/issues}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

}
\keyword{internal}
