% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3_cache_files.R
\name{recount3_cache_files}
\alias{recount3_cache_files}
\title{Locate recount3 cached files}
\usage{
recount3_cache_files(bfc = recount3_cache())
}
\arguments{
\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to, typically created by
\code{recount3_cache()}.}
}
\value{
A \code{character()} with the URLs of the recount3 files you have
downloaded.
}
\description{
This function returns the list of URLs of the recount3 files you have
stored in your \code{recount3_cache()}.
}
\examples{
## List the URLs you have downloaded
recount3_cache_files()
}
\seealso{
Other recount3 cache functions: 
\code{\link{recount3_cache_rm}()},
\code{\link{recount3_cache}()}
}
\concept{recount3 cache functions}
