% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount3_cache_rm.R
\name{recount3_cache_rm}
\alias{recount3_cache_rm}
\title{Remove recount3 cached files}
\usage{
recount3_cache_rm(bfc = recount3_cache())
}
\arguments{
\item{bfc}{A \link[BiocFileCache:BiocFileCache-class]{BiocFileCache-class}
object where the files will be cached to, typically created by
\code{recount3_cache()}.}
}
\value{
A \code{character(0)} if the removal of files was successful.
}
\description{
This function removes the recount3 files you have
stored in your \code{recount3_cache()}.
}
\examples{
## List the URLs you have downloaded
recount3_cache_files()

\dontrun{
## Now delete the cached files
recount3_cache_rm()

## List againt your recount3 files (should be empty)
recount3_cache_files()
}
}
\seealso{
Other recount3 cache functions: 
\code{\link{recount3_cache_files}()},
\code{\link{recount3_cache}()}
}
\concept{recount3 cache functions}
